move shared components to ballot_components.tsx
