+++ libs/hmpb/src/ballot_templates/nh_ballot_template.tsx
@@ -39,10 +39,6 @@ import {
   TimingMarkGrid,
   WRITE_IN_OPTION_CLASS,
   pageMarginsInches,
-} from '../ballot_components';
-import { BallotMode, PixelDimensions } from '../types';
-import { hmpbStrings } from '../hmpb_strings';
-import {
   BlankPageMessage,
   Box,
   BubbleWrapper,
@@ -53,7 +49,9 @@ import {
   Instructions,
   primaryLanguageCode,
   WriteInLabel,
-} from './vx_default_ballot_template';
+} from '../ballot_components';
+import { BallotMode, PixelDimensions } from '../types';
+import { hmpbStrings } from '../hmpb_strings';
 import { layOutInColumns } from '../layout_in_columns';
 import { Watermark } from './watermark';
